Fixed building on MSVC (#250)
* Fixed per Uniq's instructions

+++ source/nbt/Int8ArrayTag.cpp
@@ -22,7 +22,7 @@ void Int8ArrayTag::load(IDataInput& dis)
 	if (elements <= 0)
 		return;
 
-	ssize_t bytesLeft = dis.numBytesLeft();
+	long bytesLeft = dis.numBytesLeft();
 	int32_t count = elements;
 	if (bytesLeft < elements)
 		count = bytesLeft;
